requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
-e .
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.14.3
    # via -r requirements.in
black==26.3.1
    # via -r requirements.in
certifi==2026.2.25
    # via requests
cfgv==3.5.0
    # via pre-commit
charset-normalizer==3.4.7
    # via
    #   reportlab
    #   requests
click==8.3.2
    # via black
distlib==0.4.0
    # via virtualenv
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via -r requirements.in
filelock==3.25.2
    # via
    #   python-discovery
    #   virtualenv
greenlet==3.4.0
    # via sqlalchemy
identify==2.6.18
    # via pre-commit
idna==3.11
    # via
    #   email-validator
    #   requests
iniconfig==2.3.0
    # via pytest
librt==0.9.0
    # via mypy
mypy==1.20.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.10.0
    # via pre-commit
packaging==26.0
    # via
    #   black
    #   pytest
pathspec==1.0.4
    # via
    #   black
    #   mypy
pillow==12.2.0
    # via reportlab
platformdirs==4.9.6
    # via
    #   black
    #   python-discovery
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.5.1
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.in
pygments==2.20.0
    # via pytest
pytest==9.0.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-discovery==1.2.2
    # via virtualenv
python-dotenv==1.2.2
    # via pydantic-settings
pytokens==0.4.1
    # via black
pyyaml==6.0.3
    # via pre-commit
reportlab==4.4.10
    # via -r requirements.in
requests==2.33.1
    # via -r requirements.in
ruff==0.15.10
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.49
    # via -r requirements.in
structlog==25.5.0
    # via -r requirements.in
tenacity==9.1.4
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
virtualenv==21.2.1
    # via pre-commit
